Add setting for log file.
Need to use a separate log file for the webhook as it doesn't have
permissions to write to the log file. Need to reconsider how data is
stored. Should really give the different users different spaces, but
don't have a good way to write configuration files for the website user.