Try to download alternative files if first fails.
I assumed that if a file was listed by the API it would actually be
downloadable. This turns out not to be true, and for a while I was
getting continually getting a 404 error for a file even though it was
listed. So instead of selecting the best file to download in
fetch_details, return a list of files in order of preference.

This messes up caching slightly, as if the second file is used and it
has a different version, then it will still try to fetch the first file
on later iterations. This is pretty unlikely as it relies on two things
going wrong - files with different versions existing at the same time,
and the first file having a download error.