Prune deleted branches in mirror.
Updating the mirror failed because an old branch name is now used as a
directory name in one repo. I had to manually delete the file to get it
running again. I think this could still be an issue if it happens
quickly, so I'm not sure what the proper solution is.