Add ability to prioritize mirror updates
There's a race condition where the super-project is updated after a
submodule, so that the super-project includes a change, but the
submodule doesn't. It seemed fairly unlikely because of how the cron
jobs are scheduled, but it happened today, so I'm hacking in a little
fix to make the super project update first.

There isn't a general solution to this as modules can have different
submodules on different branches. But could try fetching a mirrored repo
whenever it is missing a commit.