Config option for superproject repo
To allow me to do a trial run with a different repo. This is not great.
Since most git commands just use the current remotes, when the option is
changed need to delete the local copies of the repos. Would be better to
use distinct paths and perhaps push to the url. Probably need to go
through most of the code and audit it for problems.